requirements.txt
# Core Dependencies
python-telegram-bot>=20.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
